backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lean-ai"
version = "0.1.0"
description = "Lean agentic coding assistant — plan well, give the LLM tools, let it work"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.34.0,<1.0",
    "aiosqlite>=0.20.0,<1.0",
    "pydantic>=2.10.0,<3.0",
    "pydantic-settings>=2.7.0,<3.0",
    "ollama>=0.4.0,<1.0",
    "whoosh>=2.7.4,<3.0",
    "beautifulsoup4>=4.12.0,<5.0",
    "httpx>=0.28.0,<1.0",
    "ruamel.yaml>=0.18.0,<0.19",
    "pathspec>=0.12.0,<2.0",
    "websockets>=13.0,<17.0",
    "duckduckgo-search>=7.0.0,<8.0",
    "tree-sitter>=0.24.0,<1.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    "tree-sitter-c>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-c-sharp>=0.23.0",
    "tree-sitter-php>=0.23.0",
    "tree-sitter-css>=0.23.0",
    "tree-sitter-html>=0.23.0",
]

[project.optional-dependencies]
knowledge = [
    "ebooklib>=0.18,<1.0",
    "pypdf>=5.0.0,<6.0",
    "python-docx>=1.1.0,<2.0",
]
dev = [
    "pytest>=8.0.0,<10.0",
    "pytest-asyncio>=0.24.0,<2.0",
    "ruff>=0.8.0,<1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lean_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
